+++ CHANGELOG.md
@@ -95,10 +95,10 @@
 ### Enhancements made
 
 - Update non-editable mode [#32](https://github.com/trungleduc/ipyflex/pull/32) ([@trungleduc](https://github.com/trungleduc))
-- Show or hide tabs individually  [#30](https://github.com/trungleduc/ipyflex/pull/30) ([@trungleduc](https://github.com/trungleduc))
+- Show or hide tabs individually [#30](https://github.com/trungleduc/ipyflex/pull/30) ([@trungleduc](https://github.com/trungleduc))
 - Add api to save/load template from python side. [#29](https://github.com/trungleduc/ipyflex/pull/29) ([@trungleduc](https://github.com/trungleduc))
 - Add header component [#21](https://github.com/trungleduc/ipyflex/pull/21) ([@trungleduc](https://github.com/trungleduc))
-- Create widget from factory  [#18](https://github.com/trungleduc/ipyflex/pull/18) ([@trungleduc](https://github.com/trungleduc))
+- Create widget from factory [#18](https://github.com/trungleduc/ipyflex/pull/18) ([@trungleduc](https://github.com/trungleduc))
 
 ### Maintenance and upkeep improvements